libs/result-sdk/pyproject.toml
[tool.poetry]
name = "tabula-result-sdk"
version = "1.0.0"
description = "An SDK for grading based on user-uploaded learning outcomes"
packages = [
    { include = "result_sdk" }
]

[tool.poetry.dependencies]
python = "^3.11"
tabula-common-sdk = { path = "../../libs/common-sdk", develop = true }
langgraph = "0.4.3"
pinecone = "6.0.2"
pinecone-text = "0.10.0"
pymupdf = "^1.26.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.7.0"
pillow = "^10.3.0"
requests = "^2.31.0"
opencv-python = "^4.9.0"
google-auth = "^2.29.0"
pdf2image = "^1.17.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
